# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Narvi API",
        "description": "API para la gestión de productos, usuarios, imágenes y traducciones multiidioma",
        "contact": {
            "email": "support@narvi.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/public/{lang}/products": {
            "get": {
                "tags": [
                    "Products Public"
                ],
                "summary": "Listar productos públicos",
                "description": "Endpoint público que devuelve una lista paginada de productos con traducciones aplicadas",
                "operationId": "9dcc6b5ef0c975faf61e124ce934a080",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "Código de idioma (2 letras)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de productos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de productos con traducciones aplicadas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Premium Product"
                                                    },
                                                    "url_alias": {
                                                        "type": "string",
                                                        "example": "premium-product"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Product description"
                                                    },
                                                    "images": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/public/{lang}/products/{alias}": {
            "get": {
                "tags": [
                    "Products Public"
                ],
                "summary": "Obtener producto público por alias",
                "description": "Endpoint público que devuelve un producto con las traducciones aplicadas según el idioma especificado, buscando por alias de URL",
                "operationId": "cb5e527126a7f5669df7dc78a1600259",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "Código de idioma (2 letras)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    },
                    {
                        "name": "alias",
                        "in": "path",
                        "description": "Alias de URL del producto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "premium-product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto con traducciones aplicadas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "product": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Premium Product"
                                                },
                                                "url_alias": {
                                                    "type": "string",
                                                    "example": "premium-product"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Product description"
                                                },
                                                "primary_button_url": {
                                                    "type": "string",
                                                    "example": "https://example.com/buy"
                                                },
                                                "primary_button_title": {
                                                    "type": "string",
                                                    "example": "Buy now"
                                                },
                                                "secondary_button_url": {
                                                    "type": "string",
                                                    "example": "https://example.com/info"
                                                },
                                                "secondary_button_title": {
                                                    "type": "string",
                                                    "example": "More info"
                                                },
                                                "specifications": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "images": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese su token de autenticación",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}